feat/review: created CardContent based in the mobile and desktop component view
