+++ src/routes/index.tsx
@@ -1,9 +1,10 @@
 import { Route, Routes } from "react-router-dom";
+
 import { Browse } from "../pages/Browse";
-import { Dashboard } from "../pages/Dashboard";
-import { LandingPage } from "../pages/LandingPage";
 import { SignIn } from "../pages/SignIn";
 import { SignUp } from "../pages/SignUp";
+import { Dashboard } from "../pages/Dashboard";
+import { LandingPage } from "../pages/LandingPage";
 
 import ProtectedRoute from "./protectedRoute";
 
@@ -18,6 +19,9 @@ export const MainRoutes = () => {
       </Route>
       <Route element={<ProtectedRoute />}>
         <Route path="/browse" element={<Browse />} />
+      </Route>
+      <Route element={<ProtectedRoute />}>
+        {/* <Route element={<NotFound />}> */}
       </Route>
     </Routes>
   );
